Validate that the "/web/sso_confirm" redirect URL is only used for MFA ceremonies (#56006)
* * Move ValidateClientRedirect from lib/auth/github.go to lib/client/sso/redirector.go

* Replace binary ssoTestFlow boolean with ceremony type to incorporate MFA ceremonies

* Validate that "/web/sso_confirm" is only used for MFA ceremonies

* Address comments.

* Add warning comment; Update error message.

* Update/Add tests.

* Require channel_id as exclusive query parameter for sso mfa web redirect.

* Update lib/client/sso/redirector.go


---------
