+++ lib/client/sso/redirector.go
@@ -28,10 +28,12 @@ import (
 	"net"
 	"net/http"
 	"net/http/httptest"
+	"net/netip"
 	"net/url"
 	"os"
 	"os/exec"
 	"runtime"
+	"slices"
 	"strings"
 	"time"
 
@@ -41,6 +43,7 @@ import (
 	"github.com/gravitational/teleport"
 	"github.com/gravitational/teleport/api/constants"
 	apidefaults "github.com/gravitational/teleport/api/defaults"
+	"github.com/gravitational/teleport/api/types"
 	apiutils "github.com/gravitational/teleport/api/utils"
 	"github.com/gravitational/teleport/api/utils/keys"
 	"github.com/gravitational/teleport/lib/auth/authclient"
@@ -501,4 +504,154 @@ func (rd *Redirector) wrapCallback(fn func(http.ResponseWriter, *http.Request) (
 			http.Redirect(w, r, errorURL, http.StatusFound)
 		}
 	})
+}
+
+// CeremonyType is the type of SSO ceremony.
+type CeremonyType int
+
+const (
+	// CeremonyTypeLogin ceremonies are performed during standard SSO login flows.
+	CeremonyTypeLogin CeremonyType = iota
+	// CeremonyTypeMFA ceremonies are performed during MFA flows for SSO users if SSO MFA
+	// is enabled for their origin SSO connector.
+	CeremonyTypeMFA
+	// CeremonyTypeTest ceremonies are performed by administrators with "tctl sso test".
+	CeremonyTypeTest
+)
+
+// ValidateClientRedirect validates a redirect URL provided by the client to
+// consume the SSO response at the end of the SSO Ceremony. This validation
+// differs depending on the ceremony type. Some checks are auth connector
+// specific and others depend on the ceremony type (MFA, Login, or Test).
+//
+// For Web MFA ceremonies, the redirect URL should be [WebMFARedirect] as a relative
+// path with no custom values outside of query parameters.
+//
+// For Login, Test, and non-web MFA ceremonies, the redirect URL should have:
+// - a path of "/callback"
+// - "http" scheme
+// - a hostname of "localhost", "127.0.0.1", or "::1"
+// - any port
+//
+// For non test ceremonies, If "allowed_https_hostnames" list is non-empty,
+// the redirect URL can instead have:
+// - a path of "/callback"
+// - "https" scheme
+// - a hostname that matches one in the "allowed_https_hostnames" list
+// - either empty or 443 port
+//
+// For non test ceremonies, If the "insecure_allowed_cidr_ranges" list is non-empty, URLs in both the
+// "http" and "https" schema are allowed if the hostname is an IP address that is contained in a
+// specified CIDR range on any port.
+func ValidateClientRedirect(clientRedirect string, ceremonyType CeremonyType, settings *types.SSOClientRedirectSettings) error {
+	// Warning to developers and reviewers: this validation function is critical to SSO security
+	// and any changes to it should be carefully considered from a vulnerability point of view.
+
+	if clientRedirect == "" {
+		// empty redirects are non-functional and harmless, so we allow them as
+		// they're used a lot in test code
+		return nil
+	}
+
+	u, err := url.Parse(clientRedirect)
+	if err != nil {
+		return trace.Wrap(err, "parsing client redirect URL")
+	}
+
+	if u.Opaque != "" {
+		return trace.BadParameter("unexpected opaque client redirect URL")
+	}
+	if u.User != nil {
+		return trace.BadParameter("unexpected userinfo in client redirect URL")
+	}
+	if u.Fragment != "" || u.RawFragment != "" {
+		return trace.BadParameter("unexpected fragment in client redirect URL")
+	}
+
+	// For Web MFA redirect, we expect a relative path. The proxy handling the SSO callback
+	// will redirect to itself with this relative path.
+	if u.Path == WebMFARedirect {
+		if ceremonyType != CeremonyTypeMFA {
+			return trace.BadParameter("the web mfa redirect path, \"/web/sso_confirm\", cannot be used for non-MFA flows")
+		}
+		if u.IsAbs() {
+			return trace.BadParameter("invalid client redirect URL for SSO MFA")
+		}
+		if u.Hostname() != "" {
+			return trace.BadParameter("invalid client redirect URL for SSO MFA")
+		}
+		if q, err := url.ParseQuery(u.RawQuery); err != nil {
+			return trace.Wrap(err, "parsing query in client redirect URL")
+		} else if len(q) != 1 || len(q["channel_id"]) != 1 {
+			return trace.BadParameter("malformed query parameters in client redirect URL")
+		}
+		return nil
+	}
+
+	if u.EscapedPath() != "/callback" {
+		return trace.BadParameter("invalid path in client redirect URL")
+	}
+	if q, err := url.ParseQuery(u.RawQuery); err != nil {
+		return trace.Wrap(err, "parsing query in client redirect URL")
+	} else if len(q) != 1 || len(q["secret_key"]) != 1 {
+		return trace.BadParameter("malformed query parameters in client redirect URL")
+	}
+
+	// we checked everything but u.Scheme and u.Host now
+	if u.Scheme != "http" && u.Scheme != "https" {
+		return trace.BadParameter("invalid scheme in client redirect URL")
+	}
+
+	// allow HTTP redirects to local addresses
+	allowedHTTPLocalAddrs := []string{"localhost", "127.0.0.1", "::1"}
+	if u.Scheme == "http" && slices.Contains(allowedHTTPLocalAddrs, u.Hostname()) {
+		return nil
+	}
+
+	if ceremonyType == CeremonyTypeTest {
+		return trace.AccessDenied("custom client redirect URLs are not allowed in SSO test")
+	}
+
+	const unknownRedirectHostnameErrMsg = "unknown custom client redirect URL hostname"
+	if settings == nil {
+		return trace.AccessDenied("%s", unknownRedirectHostnameErrMsg)
+	}
+
+	// allow HTTP or HTTPS redirects from IPs in specified CIDR ranges
+	hostIP, err := netip.ParseAddr(u.Hostname())
+	if err == nil {
+		hostIP = hostIP.Unmap()
+
+		for _, cidrStr := range settings.InsecureAllowedCidrRanges {
+			cidr, err := netip.ParsePrefix(cidrStr)
+			if err != nil {
+				slog.WarnContext(context.Background(), "error parsing OIDC connector CIDR prefix", "cidr", cidrStr, "err", err)
+				continue
+			}
+			if cidr.Contains(hostIP) {
+				return nil
+			}
+		}
+	}
+
+	if u.Scheme == "https" {
+		switch u.Port() {
+		default:
+			return trace.BadParameter("invalid port in client redirect URL")
+		case "", "443":
+		}
+
+		for _, expression := range settings.AllowedHttpsHostnames {
+			ok, err := utils.MatchString(u.Hostname(), expression)
+			if err != nil {
+				slog.WarnContext(context.Background(), "error compiling OIDC connector allowed HTTPS hostname regex", "regex", expression, "err", err)
+				continue
+			}
+			if ok {
+				return nil
+			}
+		}
+	}
+
+	return trace.AccessDenied("%s", unknownRedirectHostnameErrMsg)
 }
